cleaned up styling consistency, created new js module, wip
