+++ src/index.html
@@ -155,7 +155,7 @@
         </header>
         <aside class="quote">
           <blockquote>
-            <p class="quote-text">"A problem well-put is half solved."</p>
+            <p class="quote-text">“A problem well-put is half solved.”</p>
             <p class="quote-author">- John Dewey</p>
           </blockquote>
         </aside>
