+++ src/javascripts/components/projectCard/projectCard.scss
@@ -1,0 +1,1 @@
+@import "../../../styles/variables";
